+++ src/applications/vaos/new-appointment/components/TypeOfEyeCarePage.unit.spec.jsx
@@ -13,9 +13,8 @@ import {
 } from '../../tests/mocks/setup';
 
 import MockFacilityResponse from '../../tests/fixtures/MockFacilityResponse';
-import MockSchedulingConfigurationResponse, {
-  MockServiceConfiguration,
-} from '../../tests/fixtures/MockSchedulingConfigurationResponse';
+import MockSchedulingConfigurationResponse from '../../tests/fixtures/MockSchedulingConfigurationResponse';
+import MockServiceConfiguration from '../../tests/fixtures/MockServiceConfiguration';
 import {
   mockFacilitiesApi,
   mockSchedulingConfigurationsApi,
